# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: AfterHoursID API
  description: |
    Indonesian nightlife platform API - Like "Zomato for Nightlife"
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limits
    - API: 100 requests/second
    - Auth: 10 requests/second
    - Search: 20 requests/second
  version: 1.0.0
  contact:
    name: AfterHoursID Dev Team
    email: dev@afterhours.id
  license:
    name: Proprietary
servers:
  - url: https://api.afterhours.id
    description: Production server
  - url: https://staging.afterhours.id
    description: Staging server
  - url: http://localhost:3000
    description: Local development

# =====================================================
# COMPONENTS
# =====================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /api/auth/login
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [GUEST, USER, VENUE_MANAGER, ADMIN, SUPER_ADMIN]
        createdAt:
          type: string
          format: date-time
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: '#/components/schemas/User'
    
    # Venue schemas
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
...
  responses:
    Unauthorized:
      description: Unauthorized - Invalid or missing JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: unauthorized
            message: Invalid or missing authentication token
...
paths:
...
  /api/auth/login:
    post:
      summary: Login user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/me:
    get:
      summary: Get current user profile
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/refresh:
    post:
      summary: Refresh access token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/promos/checkout:
    post:
      summary: Create promo checkout
      tags:
        - Promos
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - venueId
                - title
                - description
                - tier
                - startDate
                - endDate
              properties:
                venueId:
                  type: string
                title:
                  type: string
                description:
                  type: string
                imageUrl:
                  type: string
                tier:
                  type: string
                  enum: [basic, gold, platinum]
                startDate:
                  type: string
                  format: date
                endDate:
                  type: string
                  format: date
      responses:
        '201':
          description: Checkout created
          content:
            application/json:
              schema:
                type: object
                properties:
                  promo:
                    type: object
                  payment:
                    type: object
    
  # =====================================================
  # HEALTH & METRICS
  # =====================================================
...
security:
  - bearerAuth: []